Added jwt token to verify user added jwt implementation moved hashing password to controllers folder changed todo file to update progress
