+++ app-backend/api/TODO.md
@@ -6,19 +6,18 @@ DONE create database connection
 DONE create register route
 DONE check email before registration
 DONE connect register route to database
-create login route 
-connect login route to database
-sanitize login and register route
+DONE create login route 
+DONE connect login route to database
+DONE sanitize login and register route
 create jwt tokens
+create sessions
 #flashcard section
 create example data for database
-fetch data to server
+fetch data from server
 send data to user
 create "create flashcards" route with creating flashcard to database
 create "read flashcards" route where flashcards is send to user
 send flashcards list for specific user when user logs in
-
-#ai flashcards section
 
 #addictional ideas
 after every created flashcard send it to database
