+++ app-backend/package.json
@@ -7,6 +7,7 @@
     "dotenv": "^16.4.5",
     "express": "^4.21.1",
     "expressjs": "^1.0.1",
+    "jsonwebtoken": "^9.0.2",
     "mongodb": "^6.10.0",
     "mongoose": "^8.7.2",
     "validator": "^13.12.0"
@@ -16,6 +17,7 @@
     "@types/bcrypt": "^5.0.2",
     "@types/express": "^5.0.0",
     "@types/jest": "^29.5.13",
+    "@types/jsonwebtoken": "^9.0.7",
     "@types/validator": "^13.12.2",
     "jest": "^29.7.0",
     "ts-jest": "^29.2.5"
